WIP: fixed and better tested progspace; parallel bfs
