+++ samples/infix/CMakeLists.txt
@@ -18,16 +18,25 @@ add_executable(infix_test
 target_link_libraries(infix_test
   trieste::trieste
 )
+
+add_executable(infix_progspace_test
+  progspace_test.cc
+)
+
+target_link_libraries(infix_progspace_test
+  trieste::trieste
+)
  
 add_test(NAME infix COMMAND infix_trieste test -f)
 add_test(NAME invalid_input COMMAND infix ./infix)
 set_property(TEST invalid_input PROPERTY WILL_FAIL On)
 add_test(NAME infix_test_examples COMMAND infix_test dir ${CMAKE_CURRENT_SOURCE_DIR}/examples)
 
-# depth 1 means no nested operators; op count 1 means one single assignment
-# ... this is about what might be tolerable in CI.
-# Run it yourself with bigger numbers if you want; some slight increases might be useful.
-add_test(NAME infix_test_bfs COMMAND infix_test bfs_test --op-count 1 --depth 1)
+# these params are the "useful" ones, in that they will test nested operator parsing, associativity, etc
+# they are also rather CPU hungry (~5min on my machine), so maybe this should go under a flag or something
+add_test(NAME infix_test_bfs COMMAND infix_test bfs_test --op-count 1 --depth 2)
+
+add_test(NAME infix_progspace_test COMMAND infix_progspace_test)
 
 # fuzz all configs
 add_test(
